requirements.txt
pyserial>=3.5
prompt_toolkit>=3.0.52
bleak
numpy